fix: fix the configDir failure that occured during install
* Fix the failure during the thelio install for configDir evaluation
* Eliminate a source of impurity in module evaluation

Fix the path evaluation for the config directory to account for the
module system redisgn. The evaluation of (buitins.toString ../.) was
incorrectly referring to ./modules/config instead of ./config within
the flake.

The fix: (builtins.toString ../../.)

This was needed as now modules has the following tree:
./config
./modules
  /system
    options.nix <- where this function is evaluated
  /user

The above fix ensures that the config directory is correctly found
during installation from a bootable medium.

This change has the additional side-effect of not requiring --impure
during rebuilding.